Fix HIM-2054

+++ Master/Components/Global/Include/AlarmHandler.h
@@ -83,7 +83,7 @@ public:
      * \iparam timeout = time out for alarm
      */
     /****************************************************************************/
-    void setTimeout(quint16 timeout);
+    void setTimeout(quint32 timeout);
 
     /****************************************************************************/
     /**
